TX: move static tx constructors to constructors.ts
