+++ packages/tx/src/eip4844Transaction.ts
@@ -5,16 +5,11 @@ import {
   MAX_INTEGER,
   bigIntToHex,
   bigIntToUnpaddedBytes,
-  blobsToCommitments,
-  blobsToProofs,
   bytesToBigInt,
   bytesToHex,
-  commitmentsToVersionedHashes,
   computeVersionedHash,
   equalsBytes,
-  getBlobs,
   toBytes,
-  validateNoLeadingZeroes,
 } from '@ethereumjs/util'
 
 import { BaseTransaction } from './baseTransaction.js'
@@ -23,6 +18,7 @@ import * as EIP2718 from './capabilities/eip2718.js'
 import * as EIP2930 from './capabilities/eip2930.js'
 import * as Legacy from './capabilities/legacy.js'
 import { LIMIT_BLOBS_PER_TX } from './constants.js'
+import { txFromTxData, txFromValuesArray } from './constructors.js'
 import { TransactionType } from './types.js'
 import { AccessLists, txTypeBytes } from './util.js'
 
@@ -188,41 +184,6 @@ export class BlobEIP4844Transaction extends BaseTransaction<TransactionType.Blob
     }
   }
 
-  public static fromTxData(txData: TxData, opts?: TxOptions) {
-    if (opts?.common?.customCrypto?.kzg === undefined) {
-      throw new Error(
-        'A common object with customCrypto.kzg initialized required to instantiate a 4844 blob tx'
-      )
-    }
-    const kzg = opts!.common!.customCrypto!.kzg!
-    if (txData.blobsData !== undefined) {
-      if (txData.blobs !== undefined) {
-        throw new Error('cannot have both raw blobs data and encoded blobs in constructor')
-      }
-      if (txData.kzgCommitments !== undefined) {
-        throw new Error('cannot have both raw blobs data and KZG commitments in constructor')
-      }
-      if (txData.blobVersionedHashes !== undefined) {
-        throw new Error('cannot have both raw blobs data and versioned hashes in constructor')
-      }
-      if (txData.kzgProofs !== undefined) {
-        throw new Error('cannot have both raw blobs data and KZG proofs in constructor')
-      }
-      txData.blobs = getBlobs(txData.blobsData.reduce((acc, cur) => acc + cur))
-      txData.kzgCommitments = blobsToCommitments(kzg, txData.blobs as Uint8Array[])
-      txData.blobVersionedHashes = commitmentsToVersionedHashes(
-        txData.kzgCommitments as Uint8Array[]
-      )
-      txData.kzgProofs = blobsToProofs(
-        kzg,
-        txData.blobs as Uint8Array[],
-        txData.kzgCommitments as Uint8Array[]
-      )
-    }
-
-    return new BlobEIP4844Transaction(txData, opts)
-  }
-
   /**
    * Returns the minimum of calculated priority fee (from maxFeePerGas and baseFee) and maxPriorityFeePerGas
    * @param baseFee Base fee retrieved from block
@@ -248,7 +209,7 @@ export class BlobEIP4844Transaction extends BaseTransaction<TransactionType.Blob
       )
     }
 
-    const tx = BlobEIP4844Transaction.fromTxData(
+    const tx = txFromTxData.BlobEIP4844Transaction(
       {
         ...txData,
         ...{ blobs: undefined, kzgCommitments: undefined, kzgProofs: undefined },
@@ -256,108 +217,6 @@ export class BlobEIP4844Transaction extends BaseTransaction<TransactionType.Blob
       opts
     )
     return tx
-  }
-
-  /**
-   * Instantiate a transaction from the serialized tx.
-   *
-   * Format: `0x03 || rlp([chain_id, nonce, max_priority_fee_per_gas, max_fee_per_gas, gas_limit, to, value, data,
-   * access_list, max_fee_per_data_gas, blob_versioned_hashes, y_parity, r, s])`
-   */
-  public static fromSerializedTx(serialized: Uint8Array, opts: TxOptions = {}) {
-    if (opts.common?.customCrypto?.kzg === undefined) {
-      throw new Error(
-        'A common object with customCrypto.kzg initialized required to instantiate a 4844 blob tx'
-      )
-    }
-
-    if (
-      equalsBytes(serialized.subarray(0, 1), txTypeBytes(TransactionType.BlobEIP4844)) === false
-    ) {
-      throw new Error(
-        `Invalid serialized tx input: not an EIP-4844 transaction (wrong tx type, expected: ${
-          TransactionType.BlobEIP4844
-        }, received: ${bytesToHex(serialized.subarray(0, 1))}`
-      )
-    }
-
-    const values = RLP.decode(serialized.subarray(1))
-
-    if (!Array.isArray(values)) {
-      throw new Error('Invalid serialized tx input: must be array')
-    }
-
-    return BlobEIP4844Transaction.fromValuesArray(values as TxValuesArray, opts)
-  }
-
-  /**
-   * Create a transaction from a values array.
-   *
-   * Format: `[chainId, nonce, maxPriorityFeePerGas, maxFeePerGas, gasLimit, to, value, data,
-   * accessList, signatureYParity, signatureR, signatureS]`
-   */
-  public static fromValuesArray(values: TxValuesArray, opts: TxOptions = {}) {
-    if (opts.common?.customCrypto?.kzg === undefined) {
-      throw new Error(
-        'A common object with customCrypto.kzg initialized required to instantiate a 4844 blob tx'
-      )
-    }
-
-    if (values.length !== 11 && values.length !== 14) {
-      throw new Error(
-        'Invalid EIP-4844 transaction. Only expecting 11 values (for unsigned tx) or 14 values (for signed tx).'
-      )
-    }
-
-    const [
-      chainId,
-      nonce,
-      maxPriorityFeePerGas,
-      maxFeePerGas,
-      gasLimit,
-      to,
-      value,
-      data,
-      accessList,
-      maxFeePerBlobGas,
-      blobVersionedHashes,
-      v,
-      r,
-      s,
-    ] = values
-
-    this._validateNotArray({ chainId, v })
-    validateNoLeadingZeroes({
-      nonce,
-      maxPriorityFeePerGas,
-      maxFeePerGas,
-      gasLimit,
-      value,
-      maxFeePerBlobGas,
-      v,
-      r,
-      s,
-    })
-
-    return new BlobEIP4844Transaction(
-      {
-        chainId: bytesToBigInt(chainId),
-        nonce,
-        maxPriorityFeePerGas,
-        maxFeePerGas,
-        gasLimit,
-        to,
-        value,
-        data,
-        accessList: accessList ?? [],
-        maxFeePerBlobGas,
-        blobVersionedHashes,
-        v: v !== undefined ? bytesToBigInt(v) : undefined, // EIP2930 supports v's with value 0 (empty Uint8Array)
-        r,
-        s,
-      },
-      opts
-    )
   }
 
   /**
@@ -399,7 +258,7 @@ export class BlobEIP4844Transaction extends BaseTransaction<TransactionType.Blob
       networkTxValues as BlobEIP4844NetworkValuesArray
 
     // Construct the tx but don't freeze yet, we will assign blobs etc once validated
-    const decodedTx = BlobEIP4844Transaction.fromValuesArray(txValues, { ...opts, freeze: false })
+    const decodedTx = txFromValuesArray.BlobEIP4844Transaction(txValues, { ...opts, freeze: false })
     if (decodedTx.to === undefined) {
       throw Error('BlobEIP4844Transaction can not be send without a valid `to`')
     }
@@ -608,7 +467,7 @@ export class BlobEIP4844Transaction extends BaseTransaction<TransactionType.Blob
     s = toBytes(s)
     const opts = { ...this.txOptions, common: this.common }
 
-    return BlobEIP4844Transaction.fromTxData(
+    return txFromTxData.BlobEIP4844Transaction(
       {
         chainId: this.chainId,
         nonce: this.nonce,
